+++ CHANGELOG.md
@@ -6,7 +6,7 @@ This file follows the format suggested by [Keep a CHANGELOG](https://github.com/
 
 ## [Unreleased][Unreleased]
 ### Added
-- [Feature] Add `isOptional` prop to Label React component to get properly colored optional/required labels. (#556)
+- [Feature] Add `isOptional` and `isRequired` prop to Label React component to get properly colored optional/required labels. (#556)
 
 ### Changed
 - [Patch] Use correct oui- prefix with label. (#555)